Added ADAM, adaptive momentum
The cache holds a history of squared gradients which means the bigger
the sum of the updates is, the smaller updates are made further in
training. Adam optimizer also adds a bias correction mechanism. Code is
based on RMSProp. At each iteration t, for parameters beta_1 and beta_2,
update bias, and momentum estimates, apply bias and weight corrections,
update parameters. Loss brought down to 0.073 and acc up to 98.3%